Make auth manager independent
