+++ modules/ipfs-cpinner-client/src/auth-manager/constants.ts
@@ -1,0 +1,8 @@
+// error messages
+export const NO_DID = 'No DID'
+export const NO_SIGNER = 'No Signer'
+export const NO_SERVICE_DID = 'No service DID'
+
+// storage keys
+export const ACCESS_TOKEN_KEY = 'DV_ACCESS_TOKEN'
+export const REFRESH_TOKEN_KEY = 'DV_REFRESH_TOKEN'
