+++ modules/ipfs-cpinner-client/src/auth-manager/index.ts
@@ -1,7 +1,7 @@
-import { NO_DID, NO_SIGNER, ACCESS_TOKEN_KEY, REFRESH_TOKEN_KEY, NO_SERVICE_DID } from './constants'
-import { ClientKeyValueStorage, LoginResponse, Config } from './types'
 import axios from 'axios'
 import { decodeJWT } from 'did-jwt'
+import { NO_DID, NO_SIGNER, NO_SERVICE_DID, ACCESS_TOKEN_KEY, REFRESH_TOKEN_KEY,  } from './constants'
+import { ClientKeyValueStorage, LoginResponse, Config } from './types'
 
 export default (config: Config, storage: ClientKeyValueStorage) => {
   const login = async (): Promise<LoginResponse> => {
