+++ modules/ipfs-cpinner-client/src/constants.ts
@@ -1,13 +1,7 @@
-export const ACCESS_TOKEN_KEY = 'DV_ACCESS_TOKEN'
-export const REFRESH_TOKEN_KEY = 'DV_REFRESH_TOKEN'
-
 // service errors
 export const SERVICE_MAX_STORAGE_REACHED = 'MAX_STORAGE_REACHED'
 
 // errors
-export const NO_DID = 'No DID'
-export const NO_SIGNER = 'No Signer'
-export const NO_SERVICE_DID = 'No service DID'
 export const MAX_STORAGE_REACHED = 'Max storage reached'
 export const AUTHENTICATION_ERROR = 'Authentication error'
 export const UNKNOWN_ERROR = 'Unknown error'
